requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-ai==1.0.10
pydantic-settings==2.7.1
jinja2==3.1.5

# Database
supabase==2.19.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.20.0
greenlet==3.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
email-validator==2.2.0

# Payments
stripe==11.6.0

# Email
resend==2.6.0

# AI & Content
openai==1.59.8
anthropic==0.42.0
markdown==3.7
beautifulsoup4==4.12.3
python-slugify==8.0.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
loguru==0.7.3

# Development & Testing
pytest==8.3.5
pytest-asyncio==0.25.2
black==24.10.0
ruff==0.9.2
mypy==1.14.1
redis==5.2.0
celery==5.4.0

# Monitoring
sentry-sdk[fastapi]==2.18.0
prometheus-fastapi-instrumentator==7.0.0
